+++ dart_sample/lib/sample_routes.dart
@@ -41,6 +41,7 @@ class SampleRoutes {
   static const String linearProgressIndicator = '/linear-progress-indicator';
   static const String circularProgressIndicator =
       '/circular-progress-indicator';
+  static const String refreshIndicator = '/refresh-indicator';
   static const String slider = '/slider';
   static const String rangeSlider = '/range-slider';
   static const String card = '/card';